requirements.txt
asgiref==3.8.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Django==5.0.4
django-humanize==0.1.2
django-phonenumber-field==7.3.0
Flask==3.0.3
humanize==4.9.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
phonenumbers==8.13.36
pillow==10.3.0
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
Werkzeug==3.0.2
